wandb/run-20230614_085329-1xrs4ze3/files/requirements.txt
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.7.0
appdirs==1.4.4
apprise==1.4.0
asgi-lifespan==2.1.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
blis==0.7.9
boto3==1.26.152
botocore==1.29.76
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
coolname==2.2.0
croniter==1.3.15
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.7
dateparser==1.1.8
db-dtypes==1.1.1
docker-pycreds==0.4.0
docker==6.1.3
exceptiongroup==1.1.1
fastai==2.7.12
fastapi==0.97.0
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
gitpython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.19.1
google-cloud-bigquery-storage==2.20.0
google-cloud-bigquery==3.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
griffe==0.29.0
grpcio-status==1.54.2
grpcio==1.54.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kiwisolver==1.4.4
kubernetes==26.1.0
langcodes==3.3.0
lit==16.0.5.post0
mako==1.2.4
markdown-it-py==3.0.0
markdown==3.4.3
markupsafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-boto3-s3==1.26.127
mypy-boto3-secretsmanager==1.26.135
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
orjson==3.9.1
packaging==23.1
pandas-gbq==0.19.2
pandas==1.5.2
pathspec==0.11.1
pathtools==0.1.2
pathy==0.10.1
pendulum==2.1.2
pillow==9.5.0
pip==23.1.2
prefect-aws==0.3.2
prefect-gcp==0.4.2
prefect==2.10.13
preshed==3.0.8
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
pyarrow==12.0.0
pyasn1-modules==0.3.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pydata-google-auth==1.8.0
pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
pyyaml==6.0
readchar==4.0.5
regex==2023.6.3
requests-oauthlib==1.3.1
requests==2.31.0
rich==13.4.2
rsa==4.9
s3fs==2023.6.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.25.1
setproctitle==1.3.2
setuptools==67.8.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy==3.5.3
sqlalchemy==1.4.48
srsly==2.4.6
starlette==0.27.0
sympy==1.12
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.1.0
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typer==0.7.0
typing-extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
wandb==0.15.4
wasabi==1.1.2
websocket-client==1.5.3
websockets==11.0.3
wheel==0.38.4
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0